// src/features/credit/user/components/ApplicationForms/PepsForm/types.ts
import z from 'zod'

import v from '@/utils/zodValidations'

export const publiclyExposedPersonSchema = z.object({
  organization: v.string({ required: true, max: 50 }),
  position: v.string({ required: true, max: 50 }),
  handlesPublicResources: v.boolean(),
  unbundlingDate: v.string().optional(),
})

export type PubliclyExposedPersonValues = z.infer<typeof publiclyExposedPersonSchema>

export type PepsValues = {
  isPublicServant: boolean
  publiclyExposedPerson?: PubliclyExposedPersonValues
}